Move auth redux actions into shared redux/auth slice
The session/auth actions (login, logout, refreshSession, etc.) are
app-wide concerns used by nav, metabot, and enterprise auth — not part
of the auth feature UI. Co-locate them with the existing auth reducer in
metabase/redux/auth, matching the convention used by other redux slices
(e.g. redux/user). The LoginData payload type moves alongside the login
thunk.

This removes all cross-module imports of metabase/auth/actions, which is
a prerequisite for moving auth to the feature tier (shared/feature
modules may not import from a feature module).

Pure refactor: no behavior change.
