+++ enterprise/frontend/src/metabase-enterprise/auth/index.ts
@@ -1,4 +1,3 @@
-import { LOGIN, LOGIN_GOOGLE } from "metabase/auth/actions";
 import {
   PLUGIN_AUTH_PROVIDERS,
   PLUGIN_IS_PASSWORD_USER,
@@ -6,6 +5,7 @@ import {
   PLUGIN_REDUX_MIDDLEWARES,
 } from "metabase/plugins";
 import type { AuthProvider } from "metabase/plugins/types";
+import { LOGIN, LOGIN_GOOGLE } from "metabase/redux/auth";
 import MetabaseSettings from "metabase/utils/settings";
 import { hasPremiumFeature } from "metabase-enterprise/settings";
 import type { OidcAuthProvider, User } from "metabase-types/api";
