+++ frontend/src/metabase/auth/types.ts
@@ -1,9 +1,3 @@
-export interface LoginData {
-  username: string;
-  password: string;
-  remember?: boolean;
-}
-
 export interface ForgotPasswordData {
   email: string;
 }
