+++ frontend/src/metabase/nav/containers/MainNavbar/MainNavbarContainer/MainNavbarContainer.tsx
@@ -9,7 +9,6 @@ import {
   useListDatabasesQuery,
   useReorderBookmarksMutation,
 } from "metabase/api";
-import { logout } from "metabase/auth/actions";
 import { ROOT_COLLECTION } from "metabase/collections/constants";
 import CreateCollectionModal from "metabase/collections/containers/CreateCollectionModal";
 import type { CollectionTreeItem } from "metabase/collections/utils";
@@ -22,6 +21,7 @@ import {
 import { Modal } from "metabase/common/components/Modal";
 import { PLUGIN_TENANTS } from "metabase/plugins";
 import { connect, useDispatch, useSelector } from "metabase/redux";
+import { logout } from "metabase/redux/auth";
 import type { State } from "metabase/redux/store";
 import { addUndo } from "metabase/redux/undo";
 import {
